requirements.txt
# AlphaView Portfolio Dashboard Dependencies

# Core Web Framework
dash==2.17.1
plotly==5.17.0
flask==2.3.3

# Data Processing
pandas==2.1.4
numpy==1.24.3
psycopg2-binary==2.9.9

# AWS Services
boto3==1.34.69
botocore==1.34.69

# Authentication & Security
werkzeug==2.3.7

# Excel Export Support
openpyxl==3.1.2
xlsxwriter==3.1.9

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Optional: For enhanced performance
gunicorn==21.2.0  # Production WSGI server option